requirements.old.txt
aiohttp==3.8.1
aiosignal==1.2.0
annotated-types==0.7.0
anthropic==0.39.0
anyio==4.5.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.2.2
asn1crypto==1.5.1
async-timeout==4.0.2
attrs==21.4.0
awscli==1.25.61
base58==2.1.1
bip32==3.4
bitarray==2.5.1
black==24.8.0
bloom-filter==1.3.3
boto3==1.24.60
botocore==1.27.60
certifi==2022.6.15
cffi==1.15.1
chardet==5.2.0
charset-normalizer==2.1.0
cheroot==9.0.0
CherryPy==18.7.0
click==8.1.7
coincurve==18.0.0
colorama==0.4.4
contourpy==1.1.1
cryptography==37.0.4
cycler==0.12.1
cytoolz==0.12.0
distro==1.9.0
dnspython==2.2.1
docutils==0.16
eth-abi==2.1.1
eth-account==0.5.8
eth-hash==0.4.0
eth-keyfile==0.5.1
eth-keys==0.3.4
eth-rlp==0.2.1
eth-typing==2.3.0
eth-utils==1.9.5
exceptiongroup==1.1.3
filelock==3.16.1
fonttools==4.57.0
fpdf==1.7.2
frozenlist==1.3.0
fsspec==2024.9.0
gevent==22.10.2
greenlet==2.0.2
h11==0.14.0
hexbytes==0.2.2
httpcore==1.0.6
httpx==0.27.2
huggingface-hub==0.25.1
idna==3.3
importlib-resources==5.8.0
iniconfig==2.0.0
ipfshttpclient==0.8.0a2
jaraco.classes==3.2.2
jaraco.collections==3.5.2
jaraco.context==4.1.2
jaraco.functools==3.5.1
jaraco.text==3.8.1
jellyfish==0.9.0
jiter==0.6.1
jmespath==1.0.1
jsonschema==4.7.2
kiwisolver==1.4.7
lru-dict==1.1.8
magic-admin==1.0.0
Mako==1.2.1
MarkupSafe==2.1.1
matplotlib==3.7.5
mnemonic==0.21
more-itertools==8.13.0
multiaddr==0.0.9
multidict==6.0.2
mypy-extensions==1.0.0
nameparser==1.1.3
netaddr==0.8.0
numpy==1.24.4
packaging==23.1
pandas==2.0.3
parsimonious==0.8.1
pathspec==0.12.1
pillow==10.4.0
platformdirs==4.2.2
pluggy==1.3.0
portend==3.1.0
protobuf==3.20.1
psycopg2-binary==2.9.3
py-solc-x==1.1.1
pyasn1==0.4.8
pycparser==2.21
pycryptodome==3.15.0
pydantic==2.9.2
pydantic-core==2.23.4
PyJWT==2.8.0
pymongo==4.1.1
pyparsing==3.1.4
pypdf2==3.0.1
pypng==0.20220715.0
pyrsistent==0.18.1
pytest==8.0.0
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
PyYAML==5.4.1
qrcode==7.4.2
reportlab==4.3.1
requests==2.28.1
requests-toolbelt==1.0.0
rlp==2.0.1
rsa==4.7.2
s3transfer==0.6.0
semantic-version==2.10.0
six==1.16.0
sniffio==1.3.1
stripe==9.4.0
tabulate==0.9.0
tempora==5.0.2
tokenizers==0.20.0
tomli==2.0.1
toolz==0.12.0
tqdm==4.65.0
typing-extensions==4.11.0
tzdata==2024.2
urllib3==1.26.10
us==3.2.0
varint==1.0.2
web3==5.30.0
websocket-server==0.6.4
websockets==9.1
ws4py==0.5.1
yarl==1.7.2
zc.lockfile==2.0
zipp==3.8.1
zope.event==4.6
zope.interface==6.0
